Enable NarrowCalculation error-prone check
This mostly matches IntelliJ's `integer multiplication implicitly cast
to long` inspection. While places changed do not look like bugs, so can
be viewed as false positives, this helps avoid IDE warnings and thus
makes spotting problems easier.

+++ core/trino-main/src/main/java/io/trino/operator/DirectExchangeClient.java
@@ -337,7 +337,7 @@ public class DirectExchangeClient
 
             successfulRequests++;
             // AVG_n = AVG_(n-1) * (n-1)/n + VALUE_n / n
-            averageBytesPerRequest = (long) (1.0 * averageBytesPerRequest * (successfulRequests - 1) / successfulRequests + responseSize / successfulRequests);
+            averageBytesPerRequest = (long) (1.0 * averageBytesPerRequest * (successfulRequests - 1) / successfulRequests + (double) responseSize / successfulRequests);
         }
 
         return true;
